# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NaturalLanguageQueryRequest {
  /** The natural language question to convert to SQL */
  question: string;
}

export interface QueryResult {
  /** The generated SQL query */
  sql: string;
  /** Column names in the result */
  columns: string[];
  /** Result rows as arrays of values */
  rows: (string | null)[][];
  /** Number of rows returned */
  rowCount: number;
  /**
   * Error message if query failed or was invalid
   * @nullable
   */
  error?: string | null;
  /**
   * Optional warning, e.g., empty result set
   * @nullable
   */
  warning?: string | null;
}

export interface ExplainQueryRequest {
  /** The SQL query to explain */
  sql: string;
}

export interface ExplainQueryResponse {
  /** Plain English explanation of the SQL query */
  explanation: string;
}

export interface ColumnInfo {
  name: string;
  type: string;
  nullable: boolean;
}

export interface TableInfo {
  name: string;
  columns: ColumnInfo[];
}

export interface SchemaResponse {
  /** Human-readable schema description */
  schema: string;
  tables: TableInfo[];
}

export interface SampleQuestionsResponse {
  questions: string[];
}

export interface UploadDatasetResponse {
  tableName: string;
  rowCount: number;
  columnCount: number;
  columns: string[];
  message: string;
  /** The CREATE TABLE SQL that was generated */
  sql: string;
}

export interface NewColumnDef {
  name: string;
  /** SQL type: TEXT, INTEGER, NUMERIC, BOOLEAN, DATE, TIMESTAMP, SERIAL */
  type: string;
  nullable: boolean;
  primaryKey: boolean;
  unique: boolean;
}

export interface CreateTableRequest {
  /** Name of the table to create */
  tableName: string;
  columns: NewColumnDef[];
}

export interface CreateTableResponse {
  tableName: string;
  /** The CREATE TABLE SQL that was executed */
  sql: string;
  message: string;
}

export interface DropTableResponse {
  tableName: string;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type UploadDatasetBody = {
  /** CSV or Excel (.xlsx, .xls) file to upload */
  file: Blob;
  /** Optional custom table name (auto-derived from filename if omitted) */
  tableName?: string;
};