# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sql
    description: SQL query generation and execution
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /sql/query:
    post:
      operationId: runNaturalLanguageQuery
      tags: [sql]
      summary: Convert natural language to SQL and execute it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NaturalLanguageQueryRequest"
      responses:
        "200":
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sql/explain:
    post:
      operationId: explainSqlQuery
      tags: [sql]
      summary: Explain a SQL query in plain English
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainQueryRequest"
      responses:
        "200":
          description: Explanation generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainQueryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sql/schema:
    get:
      operationId: getSchema
      tags: [sql]
      summary: Get the database schema being used
      responses:
        "200":
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaResponse"

  /sql/sample-questions:
    get:
      operationId: getSampleQuestions
      tags: [sql]
      summary: Get sample questions the user can ask
      responses:
        "200":
          description: Sample questions retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SampleQuestionsResponse"

  /sql/upload:
    post:
      operationId: uploadDataset
      tags: [sql]
      summary: Upload a CSV or Excel file and import it as a database table
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or Excel (.xlsx, .xls) file to upload
                tableName:
                  type: string
                  description: Optional custom table name (auto-derived from filename if omitted)
              required:
                - file
      responses:
        "200":
          description: Dataset uploaded and imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDatasetResponse"
        "400":
          description: Invalid file or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sql/tables:
    post:
      operationId: createTable
      tags: [sql]
      summary: Create a new database table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableRequest"
      responses:
        "201":
          description: Table created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTableResponse"
        "400":
          description: Invalid request or table already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sql/tables/{tableName}:
    delete:
      operationId: dropTable
      tags: [sql]
      summary: Drop (delete) a database table
      parameters:
        - name: tableName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table dropped successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropTableResponse"
        "400":
          description: Invalid table name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NaturalLanguageQueryRequest:
      type: object
      properties:
        question:
          type: string
          description: The natural language question to convert to SQL
      required:
        - question

    QueryResult:
      type: object
      properties:
        sql:
          type: string
          description: The generated SQL query
        columns:
          type: array
          items:
            type: string
          description: Column names in the result
        rows:
          type: array
          items:
            type: array
            items:
              type: ["string", "null"]
          description: Result rows as arrays of values
        rowCount:
          type: integer
          description: Number of rows returned
        error:
          type: ["string", "null"]
          description: Error message if query failed or was invalid
        warning:
          type: ["string", "null"]
          description: Optional warning, e.g., empty result set
      required:
        - sql
        - columns
        - rows
        - rowCount

    ExplainQueryRequest:
      type: object
      properties:
        sql:
          type: string
          description: The SQL query to explain
      required:
        - sql

    ExplainQueryResponse:
      type: object
      properties:
        explanation:
          type: string
          description: Plain English explanation of the SQL query
      required:
        - explanation

    SchemaResponse:
      type: object
      properties:
        schema:
          type: string
          description: Human-readable schema description
        tables:
          type: array
          items:
            $ref: "#/components/schemas/TableInfo"
      required:
        - schema
        - tables

    TableInfo:
      type: object
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
      required:
        - name
        - columns

    ColumnInfo:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        nullable:
          type: boolean
      required:
        - name
        - type
        - nullable

    SampleQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            type: string
      required:
        - questions

    UploadDatasetResponse:
      type: object
      properties:
        tableName:
          type: string
        rowCount:
          type: number
        columnCount:
          type: number
        columns:
          type: array
          items:
            type: string
        message:
          type: string
        sql:
          type: string
          description: The CREATE TABLE SQL that was generated
      required:
        - tableName
        - rowCount
        - columnCount
        - columns
        - message
        - sql

    CreateTableRequest:
      type: object
      properties:
        tableName:
          type: string
          description: Name of the table to create
        columns:
          type: array
          items:
            $ref: "#/components/schemas/NewColumnDef"
      required:
        - tableName
        - columns

    NewColumnDef:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          description: "SQL type: TEXT, INTEGER, NUMERIC, BOOLEAN, DATE, TIMESTAMP, SERIAL"
        nullable:
          type: boolean
        primaryKey:
          type: boolean
        unique:
          type: boolean
      required:
        - name
        - type
        - nullable
        - primaryKey
        - unique

    CreateTableResponse:
      type: object
      properties:
        tableName:
          type: string
        sql:
          type: string
          description: The CREATE TABLE SQL that was executed
        message:
          type: string
      required:
        - tableName
        - sql
        - message

    DropTableResponse:
      type: object
      properties:
        tableName:
          type: string
        message:
          type: string
      required:
        - tableName
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error