# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Convert natural language to SQL and execute it
 */
export const RunNaturalLanguageQueryBody = zod.object({
  question: zod
    .string()
    .describe("The natural language question to convert to SQL"),
});

export const RunNaturalLanguageQueryResponse = zod.object({
  sql: zod.string().describe("The generated SQL query"),
  columns: zod.array(zod.string()).describe("Column names in the result"),
  rows: zod
    .array(zod.array(zod.string().nullable()))
    .describe("Result rows as arrays of values"),
  rowCount: zod.number().describe("Number of rows returned"),
  error: zod
    .string()
    .nullish()
    .describe("Error message if query failed or was invalid"),
  warning: zod
    .string()
    .nullish()
    .describe("Optional warning, e.g., empty result set"),
});

/**
 * @summary Explain a SQL query in plain English
 */
export const ExplainSqlQueryBody = zod.object({
  sql: zod.string().describe("The SQL query to explain"),
});

export const ExplainSqlQueryResponse = zod.object({
  explanation: zod
    .string()
    .describe("Plain English explanation of the SQL query"),
});

/**
 * @summary Get the database schema being used
 */
export const GetSchemaResponse = zod.object({
  schema: zod.string().describe("Human-readable schema description"),
  tables: zod.array(
    zod.object({
      name: zod.string(),
      columns: zod.array(
        zod.object({
          name: zod.string(),
          type: zod.string(),
          nullable: zod.boolean(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Get sample questions the user can ask
 */
export const GetSampleQuestionsResponse = zod.object({
  questions: zod.array(zod.string()),
});

/**
 * @summary Upload a CSV or Excel file and import it as a database table
 */
export const UploadDatasetBody = zod.object({
  file: zod
    .instanceof(File)
    .describe("CSV or Excel (.xlsx, .xls) file to upload"),
  tableName: zod
    .string()
    .optional()
    .describe(
      "Optional custom table name (auto-derived from filename if omitted)",
    ),
});

export const UploadDatasetResponse = zod.object({
  tableName: zod.string(),
  rowCount: zod.number(),
  columnCount: zod.number(),
  columns: zod.array(zod.string()),
  message: zod.string(),
  sql: zod.string().describe("The CREATE TABLE SQL that was generated"),
});

/**
 * @summary Create a new database table
 */
export const CreateTableBody = zod.object({
  tableName: zod.string().describe("Name of the table to create"),
  columns: zod.array(
    zod.object({
      name: zod.string(),
      type: zod
        .string()
        .describe(
          "SQL type: TEXT, INTEGER, NUMERIC, BOOLEAN, DATE, TIMESTAMP, SERIAL",
        ),
      nullable: zod.boolean(),
      primaryKey: zod.boolean(),
      unique: zod.boolean(),
    }),
  ),
});

/**
 * @summary Drop (delete) a database table
 */
export const DropTableParams = zod.object({
  tableName: zod.coerce.string(),
});

export const DropTableResponse = zod.object({
  tableName: zod.string(),
  message: zod.string(),
});