# lib/api-zod/src/generated/types/uploadDatasetResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadDatasetResponse {
  tableName: string;
  rowCount: number;
  columnCount: number;
  columns: string[];
  message: string;
  /** The CREATE TABLE SQL that was generated */
  sql: string;
}